Safari-611.1.10/LayoutTests/platform/mac-catalina/fast/text/international/system-language/navigator-language/navigator-language-es-ES-expected.txt

FAIL navigator.language should be es-es. Was es.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
